requirements.txt
asgiref
dj-database-url
Django
django-filter
djangorestframework==3.15.2
whitenoise
djangorestframework-expiring-authtoken
gunicorn
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg-pool==3.2.4
pyjsparser==2.7.1
PyPrind==2.11.3
rest-framework-simplejwt
setuptools==75.6.0
typing_extensions==4.12.2
django-cors-headers
whitenoise